requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
certifi==2025.4.26
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fonttools==4.58.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jiter==0.10.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.0
openai==1.84.0
ortools==9.13.4784
packaging==24.1
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.3
prompt_toolkit==3.0.47
protobuf==6.31.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.2.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
wcwidth==0.2.13
